Move feature package into internal/controller
