+++ BUILD.bazel
@@ -76,7 +76,6 @@ filegroup(
         "//pkg/client/listers/certmanager/v1:all-srcs",
         "//pkg/controller:all-srcs",
         "//pkg/ctl:all-srcs",
-        "//pkg/feature:all-srcs",
         "//pkg/issuer:all-srcs",
         "//pkg/logs:all-srcs",
         "//pkg/metrics:all-srcs",
