+++ cmd/controller/app/BUILD.bazel
@@ -11,6 +11,7 @@ go_library(
     deps = [
         "//cmd/controller/app/options:go_default_library",
         "//cmd/util:go_default_library",
+        "//internal/controller/feature:go_default_library",
         "//pkg/acme/accounts:go_default_library",
         "//pkg/client/clientset/versioned:go_default_library",
         "//pkg/client/clientset/versioned/scheme:go_default_library",
@@ -23,7 +24,6 @@ go_library(
         "//pkg/controller/certificates/trigger:go_default_library",
         "//pkg/controller/clusterissuers:go_default_library",
         "//pkg/controller/issuers:go_default_library",
-        "//pkg/feature:go_default_library",
         "//pkg/issuer/acme:go_default_library",
         "//pkg/issuer/acme/dns/util:go_default_library",
         "//pkg/issuer/ca:go_default_library",
