+++ internal/BUILD.bazel
@@ -15,6 +15,7 @@ filegroup(
         "//internal/apis/certmanager:all-srcs",
         "//internal/apis/config/webhook:all-srcs",
         "//internal/apis/meta:all-srcs",
+        "//internal/controller/feature:all-srcs",
         "//internal/ingress:all-srcs",
         "//internal/test/paths:all-srcs",
         "//internal/vault:all-srcs",
