+++ pkg/controller/acmechallenges/BUILD.bazel
@@ -10,6 +10,7 @@ go_library(
     importpath = "github.com/jetstack/cert-manager/pkg/controller/acmechallenges",
     visibility = ["//visibility:public"],
     deps = [
+        "//internal/controller/feature:go_default_library",
         "//internal/ingress:go_default_library",
         "//pkg/acme:go_default_library",
         "//pkg/acme/accounts:go_default_library",
@@ -21,7 +22,6 @@ go_library(
         "//pkg/client/listers/certmanager/v1:go_default_library",
         "//pkg/controller:go_default_library",
         "//pkg/controller/acmechallenges/scheduler:go_default_library",
-        "//pkg/feature:go_default_library",
         "//pkg/issuer:go_default_library",
         "//pkg/issuer/acme/dns:go_default_library",
         "//pkg/issuer/acme/dns/util:go_default_library",
