+++ test/e2e/suite/conformance/certificatesigningrequests/BUILD.bazel
@@ -9,8 +9,8 @@ go_library(
     importpath = "github.com/jetstack/cert-manager/test/e2e/suite/conformance/certificatesigningrequests",
     visibility = ["//visibility:public"],
     deps = [
+        "//internal/controller/feature:go_default_library",
         "//pkg/apis/experimental/v1alpha1:go_default_library",
-        "//pkg/feature:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/feature:go_default_library",
         "//test/e2e/framework:go_default_library",
