+++ test/e2e/suite/conformance/certificatesigningrequests/suite.go
@@ -23,7 +23,7 @@ import (
 	. "github.com/onsi/ginkgo"
 	certificatesv1 "k8s.io/api/certificates/v1"
 
-	"github.com/jetstack/cert-manager/pkg/feature"
+	"github.com/jetstack/cert-manager/internal/controller/feature"
 	utilfeature "github.com/jetstack/cert-manager/pkg/util/feature"
 	"github.com/jetstack/cert-manager/test/e2e/framework"
 	"github.com/jetstack/cert-manager/test/e2e/framework/helper/featureset"
